Safari-609.1.11.1/LayoutTests/http/tests/security/protocol-compare-case-insensitive.html

<html>
<head>
    <script src="resources/cross-frame-access.js"></script>
</head>
<body>
<iframe id="aFrame" name="aFrame"></iframe>
<pre id="console"></pre>
<script>
    var url = "HTTP://127.0.0.1:8000/security/resources/cross-frame-iframe.html";
    var iframeId ="aFrame";
    var passMessage = "PASS: Protocol name compare is case insensitive!";
    var failMessage = "FAIL: Protocol name compare is case sensitive, should be case insensitive.";
    canAccessFrame(url, iframeId, passMessage, failMessage);
</script>
</body>
</html>
