third_party/blink/web_tests/platform/generic/http/tests/security/protocol-compare-case-insensitive-expected.txt

PASS: Protocol name compare is case insensitive!


--------
Frame: 'aFrame'
--------
PASS: Protocol name compare is case insensitive!

Inner iframe.
